implement depositEther and remove ether
